Adaptation à la partition /home séparée.

+++ el7/scripts/00-restaurer-profil.sh
@@ -17,19 +17,21 @@ fi
 echo
 
 for UTILISATEUR in $(ls /home); do
-  echo ":: Mise à jour du profil de l'utilisateur $UTILISATEUR."
-  rm -rf /home/$UTILISATEUR/.kde/
-  rm -rf /home/$UTILISATEUR/.local/
-  rm -rf /home/$UTILISATEUR/.winff/
-  rm -f /home/$UTILISATEUR/.gtkcdlabelrc
-  cp -R /etc/skel/.kde/ /home/$UTILISATEUR/
-  cp -R /etc/skel/.local/ /home/$UTILISATEUR/
-  cp -R /etc/skel/.winff/ /home/$UTILISATEUR/
-  cp /etc/skel/.gtkcdlabelrc /home/$UTILISATEUR/
-  chown -R $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.kde
-  chown -R $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.local
-  chown -R $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.winff
-  chown $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.gtkcdlabelrc
+  if [ $UTILISATEUR != 'lost+found' ] ; then
+    echo ":: Mise à jour du profil de l'utilisateur $UTILISATEUR."
+    rm -rf /home/$UTILISATEUR/.kde/
+    rm -rf /home/$UTILISATEUR/.local/
+    rm -rf /home/$UTILISATEUR/.winff/
+    rm -f /home/$UTILISATEUR/.gtkcdlabelrc
+    cp -R /etc/skel/.kde/ /home/$UTILISATEUR/
+    cp -R /etc/skel/.local/ /home/$UTILISATEUR/
+    cp -R /etc/skel/.winff/ /home/$UTILISATEUR/
+    cp /etc/skel/.gtkcdlabelrc /home/$UTILISATEUR/
+    chown -R $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.kde
+    chown -R $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.local
+    chown -R $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.winff
+    chown $UTILISATEUR:$UTILISATEUR /home/$UTILISATEUR/.gtkcdlabelrc
+  fi
 done
 
 echo
